+++ services/121-service/src/fsp/enum/fsp-name.enum.ts
@@ -11,5 +11,4 @@ export enum FspName {
   fspAllAttributes = 'FSP - all attributes',
   fspNoAttributes = 'FSP - no attributes',
   bankA = 'Bank A',
-  mobileMoneyProviderB = 'Mobile Money Provider B',
 }