+++ services/121-service/src/payments/fsp-integration/africas-talking/africas-talking.controller.ts
@@ -15,7 +15,7 @@ export class AfricasTalkingController {
     summary:
       'Validation callback used by Africas Talking to request validity of payment to us.',
   })
-  @ApiResponse({ status: 200, description: 'Validated' })
+  @ApiResponse({ status: 201, description: 'Validated' })
   @Post('validation')
   public async validationCallback(
     @Body() africasTalkingValidationData: AfricasTalkingValidationDto,
@@ -30,7 +30,7 @@ export class AfricasTalkingController {
     summary:
       'Notification callback used by Africas Talking to notify status of payment to us.',
   })
-  @ApiResponse({ status: 200, description: 'Notified' })
+  @ApiResponse({ status: 201, description: 'Notified' })
   @Post('notification')
   public async notificationCallback(
     @Body() africasTalkingNotificationData: AfricasTalkingNotificationDto,