+++ services/121-service/src/payments/fsp-integration/belcash/belcash.api.service.ts
@@ -1,11 +1,10 @@
-import { HttpService } from '@nestjs/axios';
 import { Injectable } from '@nestjs/common';
-import { catchError, lastValueFrom, map, of } from 'rxjs';
+import { CustomHttpService } from '../../../shared/services/custom-http.service';
 import { BelcashTransferPayload } from './belcash-transfer-payload.dto';
 
 @Injectable()
 export class BelcashApiService {
-  public constructor(private readonly httpService: HttpService) {}
+  public constructor(private readonly httpService: CustomHttpService) {}
 
   public async authenticate(): Promise<string> {
     const payload = {
@@ -14,7 +13,8 @@ export class BelcashApiService {
       token: process.env.BELCASH_API_TOKEN,
       // credentials: process.env.BELCASH_PASSWORD,
     };
-    const authenticationResult = await this.post(`authenticate`, payload);
+    const url = `${process.env.BELCASH_API_URL}/authenticate`;
+    const authenticationResult = await this.httpService.post<any>(url, payload);
     return authenticationResult.data.token;
   }
 
@@ -22,38 +22,13 @@ export class BelcashApiService {
     payload: BelcashTransferPayload,
     authorizationToken?: string,
   ): Promise<any> {
-    return await this.post('transfers', payload, authorizationToken);
-  }
-
-  private async post(
-    endpoint: string,
-    payload: any,
-    authorizationToken?: string,
-  ): Promise<any> {
-    const url = `${process.env.BELCASH_API_URL}/${endpoint}`;
-    return await lastValueFrom(
-      this.httpService
-        .post(url, payload, {
-          headers: this.createHeaders(authorizationToken),
-        })
-        .pipe(
-          map((response) => {
-            return response;
-          }),
-          catchError((err) => {
-            return of(err.response);
-          }),
-        ),
+    const headers = [
+      { name: 'Authorization', value: `Bearer ${authorizationToken}` },
+    ];
+    return await this.httpService.post(
+      `${process.env.BELCASH_API_URL}/transfers`,
+      payload,
+      headers,
     );
-  }
-
-  private createHeaders(authorizationToken?: string): object {
-    const headers = {
-      'Content-Type': 'application/json',
-    };
-    if (authorizationToken) {
-      headers['Authorization'] = `Bearer ${authorizationToken}`;
-    }
-    return headers;
   }
 }