+++ services/121-service/src/payments/fsp-integration/belcash/belcash.module.ts
@@ -1,7 +1,9 @@
 import { HttpModule } from '@nestjs/axios';
-import { Module } from '@nestjs/common';
+import { MiddlewareConsumer, Module, NestModule } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { ProgramEntity } from '../../../programs/program.entity';
+import { AzureLoggerMiddleware } from '../../../shared/middleware/azure-logger.middleware';
+import { CustomHttpService } from '../../../shared/services/custom-http.service';
 import { UserEntity } from '../../../user/user.entity';
 import { UserModule } from '../../../user/user.module';
 import { TransactionsModule } from '../../transactions/transactions.module';
@@ -17,8 +19,12 @@ import { BelcashService } from './belcash.service';
     UserModule,
     TransactionsModule,
   ],
-  providers: [BelcashService, BelcashApiService],
+  providers: [BelcashService, BelcashApiService, CustomHttpService],
   controllers: [BelcashController],
   exports: [BelcashService, BelcashApiService],
 })
-export class BelcashModule {}
+export class BelcashModule implements NestModule {
+  configure(consumer: MiddlewareConsumer): void {
+    consumer.apply(AzureLoggerMiddleware).forRoutes(BelcashController);
+  }
+}