+++ services/121-service/src/payments/transactions/dto/get-transaction.dto.ts
@@ -5,7 +5,7 @@ import { IntersolveVoucherPayoutStatus } from '../../fsp-integration/intersolve-
 
 export class GetTransactionDto {
   @ApiProperty({ example: '910c50be-f131-4b53-b06b-6506a40a2734' })
-  @Length(29, 36)
+  @Length(5, 200)
   public readonly referenceId: string;
   @ApiProperty({ example: 1 })
   @IsNotEmpty()