+++ services/121-service/src/registration/dto/set-fsp.dto.ts
@@ -4,7 +4,7 @@ import { FspName } from '../../fsp/enum/fsp-name.enum';
 
 export class SetFspDto {
   @ApiProperty({ example: '910c50be-f131-4b53-b06b-6506a40a2734' })
-  @Length(29, 36)
+  @Length(5, 200)
   public readonly referenceId: string;
   @ApiProperty({ example: 1 })
   @IsNumber()
@@ -15,7 +15,7 @@ const fspArray = Object.values(FspName).map((item) => String(item));
 
 export class UpdateChosenFspDto {
   @ApiProperty({ example: '910c50be-f131-4b53-b06b-6506a40a2734' })
-  @Length(29, 36)
+  @Length(5, 200)
   public readonly referenceId: string;
   @ApiProperty({
     enum: fspArray,