+++ services/121-service/src/registration/dto/validation-issue-data.dto.ts
@@ -25,7 +25,7 @@ export class AttributeDto {
 
 export class ValidationIssueDataDto {
   @ApiProperty({ example: '910c50be-f131-4b53-b06b-6506a40a2734' })
-  @Length(29, 36)
+  @Length(5, 200)
   public readonly referenceId: string;
   @ApiProperty({
     example: [