+++ services/121-service/src/scripts/seed-program-drc.ts
@@ -1,6 +1,5 @@
 import { Injectable } from '@nestjs/common';
 import { DataSource } from 'typeorm';
-import fspVodaCash from '../../seed-data/fsp/fsp-vodacash.json';
 import instanceDrc from '../../seed-data/instance/instance-drc.json';
 import programDrc from '../../seed-data/program/program-drc.json';
 import { InterfaceScript } from './scripts.module';
@@ -16,9 +15,6 @@ export class SeedProgramDrc implements InterfaceScript {
   public async run(): Promise<void> {
     const seedInit = await new SeedInit(this.dataSource);
     await seedInit.run();
-
-    // ***** CREATE FINANCIAL SERVICE PROVIDERS *****
-    await this.seedHelper.addFsp(fspVodaCash);
 
     // ***** CREATE PROGRAM *****
     const program = await this.seedHelper.addProgram(programDrc);