+++ services/121-service/src/scripts/seed-program-validation.ts
@@ -1,7 +1,5 @@
 import { Injectable } from '@nestjs/common';
 import { DataSource } from 'typeorm';
-import fspIntersolveNoWhatsapp from '../../seed-data/fsp/fsp-intersolve-voucher-paper.json';
-import fspIntersolve from '../../seed-data/fsp/fsp-intersolve-voucher-whatsapp.json';
 import instanceAnonymous from '../../seed-data/instance/instance-anonymous.json';
 import programValidation from '../../seed-data/program/program-validation.json';
 import { InterfaceScript } from './scripts.module';
@@ -17,10 +15,6 @@ export class SeedProgramValidation implements InterfaceScript {
   public async run(): Promise<void> {
     const seedInit = await new SeedInit(this.dataSource);
     await seedInit.run();
-
-    // ***** CREATE FINANCIAL SERVICE PROVIDERS *****
-    await this.seedHelper.addFsp(fspIntersolve);
-    await this.seedHelper.addFsp(fspIntersolveNoWhatsapp);
 
     // ***** CREATE PROGRAM *****
     const program = await this.seedHelper.addProgram(programValidation);