+++ services/121-service/src/user/user.service.ts
@@ -44,11 +44,25 @@ export class UserService {
   ) {}
 
   public async login(loginUserDto: LoginUserDto): Promise<LoginResponseDto> {
+    const saltCheck = await this.dataSource
+      .getRepository(UserEntity)
+      .createQueryBuilder('user')
+      .where({ username: loginUserDto.username })
+      .getOne();
+
+    if (!saltCheck) {
+      throw new HttpException('Unauthorized', HttpStatus.UNAUTHORIZED);
+    }
+
+    const userSalt = saltCheck.salt;
+
     const findOneOptions = {
       username: loginUserDto.username,
-      password: crypto
-        .createHmac('sha256', loginUserDto.password)
-        .digest('hex'),
+      password: userSalt
+        ? crypto
+            .pbkdf2Sync(loginUserDto.password, userSalt, 1, 32, 'sha256')
+            .toString('hex')
+        : crypto.createHmac('sha256', loginUserDto.password).digest('hex'),
     };
     const userEntity = await this.dataSource
       .getRepository(UserEntity)
@@ -211,7 +225,10 @@ export class UserService {
       ],
     });
     const updated = toUpdate;
-    updated.password = crypto.createHmac('sha256', dto.password).digest('hex');
+    updated.salt = crypto.randomBytes(16).toString('hex');
+    updated.password = crypto
+      .pbkdf2Sync(dto.password, updated.salt, 1, 32, 'sha256')
+      .toString('hex');
     await this.userRepository.save(updated);
     return await this.buildUserRO(updated);
   }